+++ transform/src/passes/replacer.ts
@@ -243,8 +243,8 @@ export class ThrowReplacer extends Visitor {
     super.visitThrowStatement(node, ref);
 
     const newThrow = Node.createIfStatement(
-      Node.createCallExpression(Node.createIdentifierExpression("isDefined", node.range), null, [Node.createPropertyAccessExpression(node.value, Node.createIdentifierExpression("rethrow", node.range), node.range)], node.range),
-      Node.createExpressionStatement(Node.createCallExpression(Node.createPropertyAccessExpression(node.value, Node.createIdentifierExpression("rethrow", node.range), node.range), null, [], node.range)),
+      Node.createCallExpression(Node.createIdentifierExpression("isDefined", node.range), null, [Node.createPropertyAccessExpression(node.value, Node.createIdentifierExpression("__try_rethrow", node.range), node.range)], node.range),
+      Node.createExpressionStatement(Node.createCallExpression(Node.createPropertyAccessExpression(node.value, Node.createIdentifierExpression("__try_rethrow", node.range), node.range), null, [], node.range)),
       Node.createThrowStatement(node.value, node.range),
       node.range,
     );